update with json and removed mic
