+++ backend/conversation.json
@@ -1,0 +1,12 @@
+[
+  {"You": "Hello Harvey, how are you doing today?"},
+  {"Other": "Hey there! I'm doing fantastic, thank you for asking. What's on your mind?"},
+  {"You": "I wanted to ask you about your approach to winning cases"},
+  {"Other": "Well, let me tell you something - I don't just win cases, I dominate them. It's all about preparation and knowing your opponent better than they know themselves."},
+  {"You": "That's interesting. Can you tell me more about your strategy?"},
+  {"Other": "Survey says... the best strategy is confidence mixed with ruthless preparation. I never go into a courtroom unless I already know I've won."},
+  {"You": "What advice would you give to someone starting their career?"},
+  {"Other": "Listen up - success isn't about luck, it's about being the smartest person in the room and making sure everyone knows it. Work harder than everyone else and never show weakness."},
+  {"You": "Thank you for the advice Harvey"},
+  {"Other": "Anytime! Remember, winners don't make excuses, they make history. That's what separates the best from the rest."}
+]
