--- backend/pretrained_models/spkrec/hyperparams.yaml
@@ -1,1 +1,0 @@
-/Users/evanferreira/.cache/huggingface/hub/models--speechbrain--spkrec-ecapa-voxceleb/snapshots/0f99f2d0ebe89ac095bcc5903c4dd8f72b367286/hyperparams.yaml
